[Misc] Moved files to their subdirectory of endtoendtests (after org/phenotips)
