+++ Readme.md
@@ -16,8 +16,8 @@ This is WIP e-2-e testing of PC.
 - cd into directory and verify `pom.xml` is there
 - Import the project on intelliJ as a Maven project.
 - Start MockMock fakeSMTP by running `java -jar MockMock.jar -p 1025 -h 8085` in the root folder. Change 1025 and 8085 to the outgoing port on PC and the port to access the email UI respectively
-- Modify `src/main/java/org.phenotips.pageobjects/BasePage.java` to specify the login credentials and the address of the PC instance (`HOMEPAGE_URL`, `ADMIN_USERNAME`, `ADMIN_PASS`, etc.).
-- Navigate to `src/main/java/org.phenotips.testcases.java` and run either `CreatePatientTest` or `LoginPageTest`. The run button is located on the class declaration line beside the line number.
+- Modify `src/main/java/org.phenotips.endtoendtests.pageobjects/BasePage.java` to specify the login credentials and the address of the PC instance (`HOMEPAGE_URL`, `ADMIN_USERNAME`, `ADMIN_PASS`, etc.).
+- Navigate to `src/main/java/org.phenotips.endtoendtests.testcases.java` and run either `CreatePatientTest` or `LoginPageTest`. The run button is located on the class declaration line beside the line number.
 	- Alternatively, right click on `MultipleClasses.xml` and then run to run both test classes
 
 ## Limitations
