+++ src/main/java/org/phenotips/endtoendtests/pageobjects/UserSignUpPage.java
@@ -1,4 +1,4 @@
-package org.phenotips.pageobjects;
+package org.phenotips.endtoendtests.pageobjects;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
