+++ src/main/java/org/phenotips/endtoendtests/testcases/AddingUsersTests.java
@@ -1,4 +1,4 @@
-package org.phenotips.testcases;
+package org.phenotips.endtoendtests.testcases;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -7,8 +7,8 @@ import java.util.List;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import org.phenotips.pageobjects.HomePage;
-import org.phenotips.pageobjects.UserSignUpPage;
+import org.phenotips.endtoendtests.pageobjects.HomePage;
+import org.phenotips.endtoendtests.pageobjects.UserSignUpPage;
 import net.bytebuddy.utility.RandomString;
 
 /**
