+++ src/main/java/org/phenotips/endtoendtests/testcases/BaseTest.java
@@ -1,6 +1,6 @@
-package org.phenotips.testcases;
+package org.phenotips.endtoendtests.testcases;
 
-import org.phenotips.pageobjects.HomePage;
+import org.phenotips.endtoendtests.pageobjects.HomePage;
 
 import java.io.File;
 import java.io.IOException;
@@ -12,7 +12,6 @@ import java.time.LocalDateTime;
 import java.time.ZonedDateTime;
 
 import org.apache.commons.io.FileUtils;
-import org.openqa.selenium.NoAlertPresentException;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.UnhandledAlertException;
