+++ src/main/java/org/phenotips/endtoendtests/testcases/CommonInfoEnums.java
@@ -1,7 +1,7 @@
-package org.phenotips.testcases;
+package org.phenotips.endtoendtests.testcases;
 
 /**
- * This class provides common enums used by both org.phenotips.pageobjects and can be used in test cases too.
+ * This class provides common enums used by both org.phenotips.endtoendtests.pageobjects and can be used in test cases too.
  * This level of indirection is actually optional for our usage. It allows for the headings
  * name to change and then we can update the enum without touching the selector.
   */
