+++ src/main/java/org/phenotips/endtoendtests/testcases/CreatePatientTest.java
@@ -1,4 +1,4 @@
-package org.phenotips.testcases;
+package org.phenotips.endtoendtests.testcases;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -7,12 +7,12 @@ import java.util.List;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import org.phenotips.pageobjects.AdminRefreshMatchesPage;
-import org.phenotips.pageobjects.CreatePatientPage;
-import org.phenotips.pageobjects.EmailUIPage;
+import org.phenotips.endtoendtests.pageobjects.AdminRefreshMatchesPage;
+import org.phenotips.endtoendtests.pageobjects.CreatePatientPage;
+import org.phenotips.endtoendtests.pageobjects.EmailUIPage;
 import net.bytebuddy.utility.RandomString;
-import org.phenotips.pageobjects.HomePage;
-import org.phenotips.pageobjects.ViewPatientPage;
+import org.phenotips.endtoendtests.pageobjects.HomePage;
+import org.phenotips.endtoendtests.pageobjects.ViewPatientPage;
 
 /**
  * Testing the creation of two very similar patients via JSON import and manually. Asserts a match at end.
