+++ src/main/java/org/phenotips/endtoendtests/testcases/xml/AllTests.xml
@@ -2,21 +2,21 @@
   <test name="FullSuite">
     <classes>
 
-      <class name="org.phenotips.SetupUsers">
+      <class name="org.phenotips.endtoendtests.SetupUsers">
         <methods>
           <include name="setEmailPort" />
           <include name="setupAutomationUsers" />
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.LoginPageTest">
+      <class name="org.phenotips.endtoendtests.testcases.LoginPageTest">
         <methods>
           <include name="loginAdminTest" />
           <include name="loginUserTest" />
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.CreatePatientTest" >
+      <class name="org.phenotips.endtoendtests.testcases.CreatePatientTest" >
         <methods>
           <include name="createPatientManually" />
           <include name="importSecondJSONPatient" />
@@ -30,7 +30,7 @@
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.AddingUsersTests" >
+      <class name="org.phenotips.endtoendtests.testcases.AddingUsersTests" >
         <methods>
           <include name="adminAddAndApproveUser" />
           <include name="userSignUpNotApproved" />
@@ -39,7 +39,7 @@
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.MatchNotificationPageTests" >
+      <class name="org.phenotips.endtoendtests.testcases.MatchNotificationPageTests" >
         <methods>
           <include name="refreshMatchesFirst" />
           <include name="matchPhenotypeOnly" />
@@ -47,7 +47,7 @@
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.PatientCreationOptionsTests" >
+      <class name="org.phenotips.endtoendtests.testcases.PatientCreationOptionsTests" >
         <methods>
           <include name="cycleThroughInfoOptions" />
           <include name="cycleThroughFamilialConditions" />
@@ -59,7 +59,7 @@
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.PedigreePageTest" >
+      <class name="org.phenotips.endtoendtests.testcases.PedigreePageTest" >
         <methods>
           <include name="basicPedigree" />
           <include name="createChild" />
@@ -69,7 +69,7 @@
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.PermissionsTests" >
+      <class name="org.phenotips.endtoendtests.testcases.PermissionsTests" >
         <methods>
           <include name="noMatchPrivatePatient" />
           <include name="cannotSeeOtherPatients" />
