+++ src/main/java/org/phenotips/endtoendtests/testcases/xml/DemoTests.xml
@@ -2,14 +2,14 @@
   <test name="TwoTests">
     <classes>
 
-      <class name="org.phenotips.SetupUsers">
+      <class name="org.phenotips.endtoendtests.SetupUsers">
         <methods>
           <include name="setEmailPort" />
           <include name="setupAutomationUsers" />
         </methods>
       </class>
 
-      <class name="org.phenotips.testcases.PatientCreationOptionsTests" >
+      <class name="org.phenotips.endtoendtests.testcases.PatientCreationOptionsTests" >
         <methods>
           <include name="cycleThroughInfoOptions" />
         </methods>
