+++ src/main/java/org/phenotips/endtoendtests/testcases/xml/MultipleClasses.xml
@@ -1,8 +1,8 @@
 <suite name="multipleClassesDemo" thread-count="1" parallel="methods" preserve-order="true">
   <test name="testngTest" thread-count="1" parallel="methods" preserve-order="true">
     <classes>
-      <class name="org.phenotips.testcases.LoginPageTest" />
-      <class name="org.phenotips.testcases.CreatePatientTest" />
+      <class name="org.phenotips.endtoendtests.testcases.LoginPageTest" />
+      <class name="org.phenotips.endtoendtests.testcases.CreatePatientTest" />
     </classes>
   </test>
 </suite>
